Версия 1.0

+++ src/main/java/Car.java
@@ -1,8 +1,8 @@
 public class Car {
     String name;
-    short speed;
+    int speed;
 
-    public Car(String name, short speed){
+    public Car(String name, int speed){
         this.name = name;
         this.speed = speed;
     }
